+++ src/detector.cc
@@ -24,7 +24,7 @@ SoundSystem& Detector::system_get()
 
 int Detector::nrj_get(int* data, int offset)
 {
-    float ret = 0.f;
+    double ret = 0.f;
     unsigned int max = offset + WINDOW;
     for (unsigned int i = offset; (i < system_.len_get()) && (i < max); i++)
         ret += ret + data[i] * data[i] / WINDOW;
@@ -35,7 +35,7 @@ void Detector::peaks_set()
 {
     for (unsigned int i = 0; i < system_.len_get() / 1024; i++)
         nrj1024_[i] = nrj_get(system_.left_data_get(), 1024 * i);
-    float sum = 0.f;
+    double sum = 0.f;
     for (unsigned int i = 0; i < 43; i++)
         sum += nrj1024_[i];
     nrj44100_[0] = sum / 43;
@@ -67,7 +67,7 @@ void Detector::laps_set()
     }
 }
 
-float Detector::bpm_get()
+double Detector::bpm_get()
 {
     laps_set();
     std::vector<int> occ(86, 0);
@@ -76,7 +76,7 @@ float Detector::bpm_get()
             occ[laps_[i]]++;
     int tmax = 0;
     int max = 0;
-    float moy = 0.f;
+    double moy = 0.f;
     for (int i = 1; i < 86; i++)
     {
         if (occ[i] > max)
@@ -88,7 +88,7 @@ float Detector::bpm_get()
     int prec = tmax - 1;
     if (occ[tmax + 1] > occ[prec])
         prec = tmax + 1;
-    float div = occ[tmax] + occ[prec];
-    moy = !div ? div : (float)(tmax * occ[tmax] + prec * occ[prec]) / div;
+    double div = occ[tmax] + occ[prec];
+    moy = !div ? div : (double)(tmax * occ[tmax] + prec * occ[prec]) / div;
     return 60.0f / (moy * (1024.f / 44100.f));
 }
